requirements.txt
fastapi
uvicorn
python-multipart

google-cloud-storage
google-cloud-firestore
google-cloud-documentai

pdf2image
pillow
